requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
asgiref==3.8.1
async-property==0.2.2
atlastk==0.13.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cloudinary==1.39.1
constantly==23.10.4
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==4.2.11
django-allauth==0.61.1
django-browser-reload==1.12.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drfasyncview==0.2.3
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iso8601==1.1.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypika-tortoise==0.1.6
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tortoise==0.1.1
tortoise-orm==0.20.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.2
